Flatten the pattern package
The class EditPattern is renamed to EditClass and is moved to the
preliminary package because there is no use outside of it. Unfortunately
it cannot be removed easily because the preliminary would need many
changes.
