+++ src/main/java/org/variantsync/diffdetective/pattern/proposed/AddWithMapping.java
@@ -1,8 +1,8 @@
-package org.variantsync.diffdetective.pattern.elementary.proposed;
+package org.variantsync.diffdetective.pattern.proposed;
 
 import org.variantsync.diffdetective.diff.difftree.DiffNode;
 import org.variantsync.diffdetective.diff.difftree.DiffType;
-import org.variantsync.diffdetective.pattern.elementary.ElementaryPattern;
+import org.variantsync.diffdetective.pattern.ElementaryPattern;
 
 /**
  * Our AddWithMapping pattern from the ESEC/FSE'22 paper.
