+++ src/main/java/org/variantsync/diffdetective/pattern/proposed/Specialization.java
@@ -1,24 +1,24 @@
-package org.variantsync.diffdetective.pattern.elementary.proposed;
+package org.variantsync.diffdetective.pattern.proposed;
 
 import org.prop4j.Node;
 import org.variantsync.diffdetective.analysis.logic.SAT;
 import org.variantsync.diffdetective.diff.difftree.DiffNode;
 import org.variantsync.diffdetective.diff.difftree.DiffType;
-import org.variantsync.diffdetective.pattern.elementary.ElementaryPattern;
+import org.variantsync.diffdetective.pattern.ElementaryPattern;
 
 /**
- * Our Generalization pattern from the ESEC/FSE'22 paper.
+ * Our Specialization pattern from the ESEC/FSE'22 paper.
  * @author Paul Bittner, Sören Viegener
  */
-final class Generalization extends ElementaryPattern {
-    Generalization() {
-        super("Generalization", DiffType.NON);
+final class Specialization extends ElementaryPattern {
+    Specialization() {
+        super("Specialization", DiffType.NON);
     }
 
     @Override
     protected boolean matchesArtifactNode(DiffNode artifactNode) {
         final Node pcb = artifactNode.getBeforeFeatureMapping();
         final Node pca = artifactNode.getAfterFeatureMapping();
-        return SAT.implies(pcb, pca) && !SAT.implies(pca, pcb);
+        return !SAT.implies(pcb, pca) && SAT.implies(pca, pcb);
     }
 }
