+++ src/main/java/org/variantsync/diffdetective/preliminary/pattern/elementary/FeatureContextOfRemFromPC.java
@@ -3,8 +3,8 @@ package org.variantsync.diffdetective.preliminary.pattern.elementary;
 import org.prop4j.Node;
 import org.variantsync.diffdetective.diff.Lines;
 import org.variantsync.diffdetective.diff.difftree.DiffNode;
-import org.variantsync.diffdetective.pattern.EditPattern;
-import org.variantsync.diffdetective.pattern.elementary.proposed.ProposedElementaryPatterns;
+import org.variantsync.diffdetective.preliminary.pattern.Pattern;
+import org.variantsync.diffdetective.pattern.proposed.ProposedElementaryPatterns;
 import org.variantsync.diffdetective.preliminary.analysis.data.PatternMatch;
 import org.variantsync.diffdetective.preliminary.evaluation.FeatureContext;
 import org.variantsync.diffdetective.preliminary.pattern.FeatureContextReverseEngineering;
@@ -12,7 +12,7 @@ import org.variantsync.diffdetective.preliminary.pattern.FeatureContextReverseEn
 @Deprecated
 public final class FeatureContextOfRemFromPC implements FeatureContextReverseEngineering<DiffNode> {
     @Override
-    public EditPattern<DiffNode> getPattern() {
+    public Pattern<DiffNode> getPattern() {
         return ProposedElementaryPatterns.RemFromPC;
     }
